refactor: lift AppShell into app/(admin)/layout.tsx route group
The admin app's sidebar nav flashed the employee-only "My Timesheet"
view for ~200ms on every navigation, then expanded to the full nav.
Root cause: every page.tsx wrapped its content in <AppShell>, so on
each navigation the shell remounted, useUserRole() restarted, and
during the checkAdminStatus round-trip isAdmin defaulted to false →
employee-only nav rendered briefly.

The conventional Next.js App Router fix is a route group with its own
layout.tsx. Layout.tsx files are preserved across navigations within
the segment — only `children` swaps, like an Angular router-outlet.

Changes:
- New app/(admin)/layout.tsx wraps children with AppShell + the
  Admin/Employee guard switch by path. Lives at the route-group level
  so AppShell stays mounted across navigations and useUserRole's
  state survives.
- 19 admin route directories + the root home page move into
  app/(admin)/. Route groups don't affect URLs, so /users, /timesheet,
  /, etc. still serve at the same paths. (api/, login/, sign/, and
  apple-pay-checkout/ stay at the app root because they're either
  framework routes or public.)
- Every admin page.tsx loses its <AppShell> wrapper (replaced with
  fragments) and its AppShell import. No page passes custom AppShell
  props, so this was mechanical. Relative imports gain one extra
  ../ segment to reflect the deeper nesting.
- auth-guard-wrapper.tsx simplifies to just AuthGuard for sign-in;
  the role-guard switch moved into the (admin) layout.

This also restores the EmployeeGuard wiring for /timesheet that was
intended in the timekeeping MVP — currently main applies AdminGuard
to /timesheet, which would block employees who tried to log hours.
